+++ lib/pushi/__init__.py
@@ -29,6 +29,7 @@ from . import apn
 from . import app
 from . import base
 from . import event
+from . import messaging
 from . import smtp
 from . import subscription
 from . import web
@@ -38,6 +39,7 @@ from .apn import APNAPI
 from .app import AppAPI
 from .base import BASE_URL, BASE_WS_URL, API
 from .event import EventAPI
+from .messaging import MessagingAPI
 from .smtp import SMTPAPI
 from .subscription import SubscriptionAPI
 from .web import WebAPI